fix#2: log_profile
