+++ app/Repositories/LogProfileRepository.php
@@ -1,7 +1,7 @@
 <?php
 namespace App\Repositories;
 
-use App\Models\LogProfile;
+use App\Models\ProfileLog;
 use Carbon\Carbon;
 use Auth;
 
@@ -13,14 +13,14 @@ class LogProfileRepository
         return Carbon::now()->toDateTimeString();
     }
 
-    private function getCurrentAdmin() {
+    private function getCurrentUser() {
         return Auth::user();
     }
 
     public function createLog($userId, array $fieldDiff) {
-        LogProfile::create([
+        ProfileLog::create([
             'time_change' => $this->getCurrentTimeStamp(),
-            'admin_name' => $this->getCurrentAdmin()->name,
+            'update_user_id' => $this->getCurrentUser()->id,
             'field_change' => implode(", ", $fieldDiff),
             'action' => __('profile.edit_action'),
             'user_id' => $userId,
@@ -28,6 +28,6 @@ class LogProfileRepository
     }
 
     public function getAllLogs() {
-        return LogProfile::orderBy('created_at', 'desc')->paginate(self::NUMBER_OF_ITEM);
+        return ProfileLog::orderBy('created_at', 'desc')->paginate(self::NUMBER_OF_ITEM);
     }
 }
