+++ app/Repositories/ProfileRepository.php
@@ -42,17 +42,4 @@ class ProfileRepository implements ProfileInterface
             'fullname' => $user->name,
         ]);
     }
-
-    public function getFieldDiff($request, $profile) {
-        $fieldDiff = [];
-        $fields = array_intersect(array_keys($request), $profile->getFillable());
-        
-        foreach($fields as $field) {
-            if ($profile->$field != $request[$field]) {
-                $fieldDiff[] = $field;
-            }
-        }
-
-        return $fieldDiff;
-    }
 }
